cleaned up an issue and forked to add functiion
cleaned up an issue and forked ACC-2CH to add functionality.  Now, we have ffmpeg_Eng-Audio-to-ACC-2CH_MV, which adds a cp and rm command for the original file.  It copies the original file to the temp directory for the conversion, helping to prevent a stall on larger files, and then deletes the file in the temp directory.  The original file in the target directory remains un-changed.